+++ src/common.h
@@ -1,4 +1,6 @@
 void printVersion(char* programName);
 void printMenu(int line, int col, char *menustring);
 void topLineMessage(const char *message);
+int readline(char *buffer, int buflen, char *oldbuf);
+void clear_workspace();
 int exittoshell();
