--- src/showviews.h
@@ -1,4 +1,0 @@
-int directory_view(char * currentpwd);
-char *getcwd(char *buf, size_t size);
-int global_menu();
-void clear_workspace();
